Principal Component Analysis
Thus, PCA is a method that brings together:
A measure of how each variable is associated with one another. (Covariance matrix.)
The directions in which our data are dispersed. (Eigenvectors.)
The relative importance of these different directions. (Eigenvalues.)
PCA combines our predictors and allows us to drop the eigenvectors that are relatively unimportant.
Are there extensions to PCA?


ref:https://towardsdatascience.com/a-one-stop-shop-for-principal-component-analysis-5582fb7e0a9c